refactor(realworld): divide common to peripherals and wrappers
