+++ rlinf/envs/realworld/dosw1/tasks/__init__.py
@@ -19,8 +19,8 @@ from typing import Any, Mapping
 import gymnasium as gym
 from gymnasium.envs.registration import register
 
-from rlinf.envs.realworld.common.wrappers import LeaderFollowerKeyboardIntervention
 from rlinf.envs.realworld.dosw1.tasks.pick import PickEnv as PickEnv
+from rlinf.envs.realworld.wrappers import LeaderFollowerKeyboardIntervention
 
 
 def _maybe_apply_keyboard_intervention(
