+++ rlinf/envs/realworld/peripherals/teleop/keyboard/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2025 The RLinf Authors.
+# Copyright 2026 The RLinf Authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,7 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from rlinf.envs.realworld.peripherals.hands.ruiyan_hand import RuiyanHand
-
-__all__ = ["RuiyanHand"]
