+++ rlinf/envs/realworld/wrappers/reward_done_wrapper.py
@@ -17,7 +17,9 @@ from typing import Any, SupportsFloat
 import gymnasium as gym
 from gymnasium.core import ActType, ObsType
 
-from rlinf.envs.realworld.common.keyboard.keyboard_listener import KeyboardListener
+from rlinf.envs.realworld.peripherals.teleop.keyboard.keyboard_listener import (
+    KeyboardListener,
+)
 
 
 class BaseKeyboardRewardDoneWrapper(gym.Wrapper):
