+++ rlinf/envs/realworld/wrappers/spacemouse_intervention.py
@@ -17,7 +17,9 @@ import time
 import gymnasium as gym
 import numpy as np
 
-from rlinf.envs.realworld.common.spacemouse.spacemouse_expert import SpaceMouseExpert
+from rlinf.envs.realworld.peripherals.teleop.spacemouse.spacemouse_expert import (
+    SpaceMouseExpert,
+)
 
 
 def sample_gripper_action(is_open: bool) -> np.ndarray:
